plan/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
charset-normalizer==2.1.1
decorator==5.1.1
discord==2.1.0
discord.py==2.1.0
Django==4.1.2
django-extensions==3.2.1
django-factory-boy==1.0.0
django-filter==22.1
django-guardian==2.4.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.3
frozenlist==1.3.3
guardian==0.2.1
idna==3.4
iniconfig==1.1.1
ipython==8.6.0
jedi==0.18.1
Jinja2==2.11.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
model-bakery==1.9.0
multidict==6.0.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.5
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
tomli==2.0.1
traitlets==5.5.0
wcwidth==0.2.5
yarl==1.8.1
